(cleanup/fix): ImplementWebhooks method should only be used by the e2e tests and should be under its package
